Split utils/index.js into modular structure
