+++ tests/helpers/dom.test.js
@@ -3,7 +3,7 @@
  */
 
 import { beforeEach, describe, expect, it } from 'vitest'
-import DOMHelper from '@/utils/dom.js'
+import DOMHelper from '@/helpers/dom.js'
 
 // Setup test environment
 beforeEach(() => {
